Split the dummy-resources in an engine
