+++ decidim-dev/app/events/decidim/dummy_resources/dummy_resource_event.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+module Decidim
+  module DummyResources
+    class DummyResourceEvent < Events::BaseEvent
+      include Decidim::Events::EmailEvent
+      include Decidim::Events::NotificationEvent
+    end
+  end
+end
