+++ decidim-dev/app/jobs/decidim/dummy_resources/hide_all_created_by_author_job.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Decidim
+  module DummyResources
+    class HideAllCreatedByAuthorJob < Decidim::HideAllCreatedByAuthorJob
+      protected
+
+      def base_query
+        Decidim::DummyResources::DummyResource.not_hidden.where(author:)
+      end
+    end
+  end
+end
