+++ decidim-dev/app/models/decidim/dummy_resources/application_record.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module Decidim
+  module DummyResources
+    class ApplicationRecord < ActiveRecord::Base
+      self.abstract_class = true
+    end
+  end
+end
