+++ decidim-dev/app/models/decidim/dummy_resources/coauthorable_dummy_resource.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+module Decidim
+  module DummyResources
+    class CoauthorableDummyResource < ApplicationRecord
+      include ::Decidim::Coauthorable
+      include HasComponent
+    end
+  end
+end
