+++ decidim-dev/app/models/decidim/dummy_resources/nested_dummy_resource.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+module Decidim
+  module DummyResources
+    class NestedDummyResource < ApplicationRecord
+      include Decidim::Resourceable
+      belongs_to :dummy_resource
+    end
+  end
+end
