+++ decidim-dev/lib/decidim/dummy_resources/admin.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+module Decidim
+  module DummyResources
+    # This module contains all the domain logic associated to Decidim's Budgets
+    # component admin panel.
+    module Admin
+    end
+  end
+end
